update PhoneNumberViewController's RegistrationTextFieldDelegate methods
