+++ Wire-iOS/Sources/Authentication/RegistrationPersonal/PhoneNumberViewController+PastePhoneNumber.swift
@@ -21,16 +21,50 @@ import Foundation
 extension PhoneNumberViewController {
     @objc
     @discardableResult
-    func pastePhoneNumber(_ phoneNumber: NSString?) -> Bool {
+    func pastePhoneNumber(_ phoneNumber: String?) -> Bool {
         guard let phoneNumber = phoneNumber else { return false }
 
         return phoneNumber.shouldPasteAsPhoneNumber(presetCountry: country){ country, phoneNumber in
             if let country = country, let phoneNumber = phoneNumber {
-                self.setCountry(country)
+                self.country = country
 
                 self.phoneNumberField.text = phoneNumber;
                 self.updateRightAccessory(forPhoneNumber: phoneNumber)
             }
         }
     }
+}
+
+extension PhoneNumberViewController: RegistrationTextFieldDelegate {
+    public func textField(_ textField: UITextField?, shouldPasteCharactersIn range: NSRange, replacementString string: String?) -> Bool {
+        return pastePhoneNumber(string)
+    }
+
+    @objc(textField:shouldChangeCharactersInRange:replacementString:)
+    public func textField(_ textField: UITextField, shouldChangeCharactersIn range: NSRange, replacementString string: String) -> Bool {
+        guard let newString = (textField.text as NSString?)?.replacingCharacters(in: range, with: string) else { return false }
+
+        guard let country = country else { return true }
+
+        ///If the textField is empty and a replacementString with longer than 1 char, it is likely to insert from autoFill.
+        if textField.text?.count == 0 && string.count > 1 {
+            return pastePhoneNumber(string)
+        }
+
+
+        let number = PhoneNumber(countryCode: country.e164.uintValue, numberWithoutCode: newString)
+
+        switch number.validate() {
+        case .containsInvalidCharacters, .tooLong:
+            return false
+        case .tooShort:
+            break
+        default:
+            break
+        }
+
+        updateRightAccessory(forPhoneNumber: phoneNumber)
+        return true
+    }
+
 }
